moved landing screen
