+++ apolloschurchapp/src/ui/LandingScreen/LandingScreen.tests.js
@@ -4,7 +4,7 @@ import { GradientOverlayImage } from '@apollosproject/ui-kit';
 
 import Providers from '../../Providers';
 
-import LandingScreen from '.';
+import LandingScreen from './LandingScreen';
 
 describe('The Onboarding LandingScreen component', () => {
   it('should render', () => {
